[tools] perf/traceevent: Check for spaces in character array
Message-id: <1408462094-14194-262-git-send-email-jolsa@redhat.com>
Patchwork-id: 88283
O-Subject: [PATCH RHEL7.1 BZ1131394 261/325] tools lib traceevent: Check for spaces in character array
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit 5efb9fbd5f1bfe4435bd0a3ea5f0e187875509c2
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Fri Nov 1 17:53:58 2013 -0400

description
===========
Currently when using the raw format for fields, when looking at a
character array, to determine if it is a string or not, we make sure all
characters are "isprint()". If not, then we consider it a numeric array,
and print the hex numbers of the characters instead.

But it seems that '\n' fails the isprint() check! Add isspace() to the
check as well, such that if all characters pass isprint() or isspace()
it will assume the character array is a string.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>