[tools] perf/fs: Rename NAME_find_mountpoint() to NAME__mountpoint()
Message-id: <1408462094-14194-271-git-send-email-jolsa@redhat.com>
Patchwork-id: 88232
O-Subject: [PATCH RHEL7.1 BZ1131394 270/325] perf fs: Rename NAME_find_mountpoint() to NAME__mountpoint()
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit cf38fadade52df937521dd70d4437df1a9354cd9
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Nov 5 14:48:50 2013 -0300

description
===========
Shorten it, "finding" it is an implementation detail, what callers want
is the pathname, not to ask for it to _always_ do the lookup.

And the existing implementation already caches it, i.e. it doesn't
"finds" it on every call.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>