[tools] perf/trace: Beautify fifth argument of mmap() as fd
Message-id: <1408462094-14194-293-git-send-email-jolsa@redhat.com>
Patchwork-id: 88309
O-Subject: [PATCH RHEL7.1 BZ1131394 292/325] perf trace: Beautify fifth argument of mmap() as fd
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit 73faab3a421351c342b10b0177fec9eea1a1d809
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Nov 12 15:24:59 2013 +0900

description
===========
The fifth argument of mmap syscall is fd and it often contains -1 as a
value for anon mappings.  Without this patch it doesn't show the file
name as well as it shows -1 as 4294967295.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>