[tools] perf/traceevent: Fix use of multiple options in processing field
Message-id: <1408462094-14194-324-git-send-email-jolsa@redhat.com>
Patchwork-id: 88258
O-Subject: [PATCH RHEL7.1 BZ1131394 323/325] tools lib traceevent: Fix use of multiple options in processing field
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit eff2c92f86c2ac2a0eab3749d58be39592293c3a
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Mon Nov 18 14:23:14 2013 -0500

description
===========
Jiri Olsa reported that the scsi_dispatch_cmd_done event failed to parse
with:

  Error: expected type 5 but read 4
  Error: expected type 5 but read 4

The problem is with this part of the print_fmt:

  __print_symbolic(((REC->result) >> 24) & 0xff, ...

The __print_symbolic() helper function's first parameter is the field to
use to determine what symbol to print based on the value of the result.
The parser can handle one operation, but it can not handle multiple
operations ('>>' and '&').

Add code to process all operations for the field argument for
__print_symbolic() as well as __print_flags().
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>