[crypto] pkcs7: Find the right key in the PKCS#7 key list and verify the signature
Message-id: <1408485785-777-8-git-send-email-vgoyal@redhat.com>
Patchwork-id: 88367
O-Subject: [RHEL7.1 PATCH 07/33] PKCS#7: Find the right key in the PKCS#7 key list and verify the signature
Bugzilla: 1129779
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ:1129779

commit a4730357ee724f8c64f0292541ba3da8a95510fb
Author: David Howells <dhowells@redhat.com>
Date: Tue, 1 Jul 2014 16:40:19 +0100

Find the appropriate key in the PKCS#7 key list and verify the signature with
it.  There may be several keys in there forming a chain.  Any link in that
chain or the root of that chain may be in our keyrings.

Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>