[crypto] pefile: Strip the wrapper off of the cert data block
Message-id: <1408485785-777-15-git-send-email-vgoyal@redhat.com>
Patchwork-id: 88363
O-Subject: [RHEL7.1 PATCH 14/33] pefile: Strip the wrapper off of the cert data block
Bugzilla: 1129779
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ:1129779

commit 09dacbbda935895a4898e517e1248c11ca493216
Author: David Howells <dhowells@redhat.com>
Date: Tue, 1 Jul 2014 16:02:51 +0100

The certificate data block in a PE binary has a wrapper around the PKCS#7
signature we actually want to get at.  Strip this off and check that we've got
something that appears to be a PKCS#7 signature.

Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>