[crypto] pkcs7: X.509 certificate issuer and subject are mandatory fields in the ASN.1
Message-id: <1408485785-777-32-git-send-email-vgoyal@redhat.com>
Patchwork-id: 88385
O-Subject: [RHEL7.1 PATCH 31/33] PKCS#7: X.509 certificate issuer and subject are mandatory fields in the ASN.1
Bugzilla: 1129779
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ:1129779

commit 412eccbadfbb1521e62ae53db57d782d6bc36993
Author: David Howells <dhowells@redhat.com>
Date: Thu, 31 Jul 2014 14:46:44 +0100

X.509 certificate issuer and subject fields are mandatory fields in the ASN.1
and so their existence needn't be tested for.  They are guaranteed to end up
with an empty string if the name material has nothing we can use (see
x509_fabricate_name()).

Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>