[fs] gfs2: Request demote when a "try" flock fails
Message-id: <1951528458.10177163.1408710321908.JavaMail.zimbra@redhat.com>
Patchwork-id: 88705
O-Subject: [RHEL7.1 PATCH][bz#1132154] GFS2: Request demote when a "try" flock fails
Bugzilla: 1132154
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

I previously posted this for RHEL6.6. This is the RHEL7.1 port.
The upstream patch is here:

http://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/fs/gfs2?id=2ddfbdd6848d496d8088b28992d257fd02e58c9d

I tested it with my four-node cluster using reliable recreation scenarios.

Patch description:
This patch changes the flock code so that it uses the TRY_1CB flag
instead of the TRY flag on the first attempt. That forces any holding
nodes to issue a dlm callback, which requests a demote of the glock.
Then, if the "try" failed, it sleeps a small amount of time for the
demote to occur. Then it tries again, for an increasing amount of time.
Subsequent attempts to gain the "try" lock don't use "_1CB" so that
only one callback is issued.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>

rhbz#1132154

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>