[tools] perf/kvm: Fix spurious '=' use in man page
Message-id: <1408953210-26343-74-git-send-email-jolsa@redhat.com>
Patchwork-id: 88802
O-Subject: [PATCH RHEL7.1 BZ1133083 073/322] perf kvm: Fix spurious '=' use in man page
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit 316bd98a9aa47693215e0d3a5bcc83f51691ba93
Author: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>
Date: Wed Dec 4 17:56:43 2013 -0500

description
===========
As option --host and --guest request no input for it, there should not
be a '=' after them in the man page sources.

And --output expects a filename as the input, so there should be a '='
after it.

This patch removes the needless '=' after --guest and --host, and adds a
'=' after --output in perf-kvm.txt.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>