[tools] perf/symbols: Constify dso->long_name
Message-id: <1408953210-26343-132-git-send-email-jolsa@redhat.com>
Patchwork-id: 88948
O-Subject: [PATCH RHEL7.1 BZ1133083 131/322] perf symbols: Constify dso->long_name
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit bf4414ae7b86cddca60a5b510954a37d30583a1f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Dec 10 15:19:23 2013 -0300

description
===========
Same reason as for dso->short_name, it may point to a const string, and
in most places it is treated as const, i.e. it is just accessed for
using its contents as a key or to show it on reports.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>