[tools] perf/evlist: Send the errno in the signal when workload fails
Message-id: <1408953210-26343-218-git-send-email-jolsa@redhat.com>
Patchwork-id: 88887
O-Subject: [PATCH RHEL7.1 BZ1133083 217/322] perf evlist: Send the errno in the signal when workload fails
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit f33cbe72e6166b97d6fa2400cb00a885b47999d7
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Jan 2 15:11:25 2014 -0300

description
===========
When a tool uses perf_evlist__start_workload and the supplied workload
fails (e.g.: its binary wasn't found), perror was being used to print
the error reason.

This is undesirable, as the caller may be a GUI, when it wants to have
total control of the error reporting process.

So move to using sigaction(SA_SIGINFO) + siginfo_t->sa_value->sival_int
to communicate to the caller the errno and let it print it using the UI
of its choosing.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>