[tools] perf/callchain: Spare double comparison of callchain first entry
Message-id: <1408953210-26343-285-git-send-email-jolsa@redhat.com>
Patchwork-id: 88998
O-Subject: [PATCH RHEL7.1 BZ1133083 284/322] perf callchain: Spare double comparison of callchain first entry
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit b965bb41061ad8d3eafda6e7feef89279fcd3916
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Tue Jan 14 16:37:15 2014 +0100

description
===========
When a new callchain child branch matches an existing one in the rbtree,
the comparison of its first entry is performed twice:

1) From append_chain_children() on branch lookup

2) If 1) reports a match, append_chain() then compares all entries of
the new branch against the matching node in the rbtree, and this
comparison includes the first entry of the new branch again.

Lets shortcut this by performing the whole comparison only from
append_chain() which then returns the result of the comparison between
the first entry of the new branch and the iterating node in the rbtree.
If the first entry matches, the lookup on the current level of siblings
stops and propagates to the children of the matching nodes.

This results in less comparisons performed by the CPU.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>