[tools] perf: Fix traceevent plugin path definitions
Message-id: <1408953210-26343-295-git-send-email-jolsa@redhat.com>
Patchwork-id: 89005
O-Subject: [PATCH RHEL7.1 BZ1133083 294/322] perf tools: Fix traceevent plugin path definitions
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit b935a58dbff457c27fd63e1e1bb29db20b2ee6a8
Author: Josh Boyer <jwboyer@fedoraproject.org>
Date: Wed Jan 22 10:01:48 2014 -0500

description
===========
The plugindir_SQ definition contains $(prefix) which is not needed as
the $(libdir) definition already contains prefix in it.  This leads to
the path including an extra prefix in it, e.g. /usr/usr/lib64.

The -DPLUGIN_DIR defintion includes DESTDIR.  This is incorrect, as it
sets the plugin search path to include the value of DESTDIR.  DESTDIR is
a mechanism to install in a non-standard location such as a chroot or an
RPM build root.  In the RPM case, this leads to the search path being
incorrect after the resulting RPM is installed (or in some cases an RPM
build failure).

Remove both of these unnecessary inclusions.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>