[tools] perf/annotate: Check availability of annotate when processing samples
Message-id: <1408953210-26343-314-git-send-email-jolsa@redhat.com>
Patchwork-id: 88945
O-Subject: [PATCH RHEL7.1 BZ1133083 313/322] perf annotate: Check availability of annotate when processing samples
Bugzilla: 1133083
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1133083
https://bugzilla.redhat.com/show_bug.cgi?id=1133083

upstream
========
commit 48c65bda95d692076de7e5eae3188ddae8635dca
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Feb 20 10:32:53 2014 +0900

description
===========
The TUI of perf report and top support annotation, but stdio and GTK
don't.  So it should be checked before calling hist_entry__inc_addr_
samples() to avoid wasting resources that will never be used.

perf annotate need it regardless of UI and sort keys, so the check
of whether to allocate resources should be on the tools that have
annotate as an option in the TUI, 'report' and 'top', not on the
function called by all of them.

It caused perf annotate on ppc64 to produce zero output, since the
buckets were not being allocated.

[ Renamed (report,top)__needs_annotate() to ui__has_annotation() ]
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>