[ethernet] i40e/i40evf: fix TSO accounting
Message-id: <1409658042-22689-156-git-send-email-sassmann@redhat.com>
Patchwork-id: 90354
O-Subject: [RHEL7.1 PATCH 155/281] i40e/i40evf: fix TSO accounting
Bugzilla: 1091126
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jesse Brandeburg <jesse.brandeburg@intel.com>

The TSO logic in the transmit path had some assumptions that
have been broken now that the kernel can send as much as 32kB
in a single skb->frag[.] entry, even on a system with 4kB pages.

This fixes the assumptions and allows the kernel to operate
as efficiently as possible with both SENDFILE and SEND.

In addition, the hardware limit of data contained in a descriptor is
changed to the next power of two below where it currently is in
order to align to a power of two value, preventing a single byte
of data in a descriptor.

Change-ID: I6af1f0b87c1458e10644dbd47541591075a52651
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit 980093ebf777f95459eabcf0ca7a29db7e9a10f7)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>