[ethernet] i40e: Allow RSS table entry range and GPS to be any number, not necessarily power of 2
Message-id: <1409658042-22689-170-git-send-email-sassmann@redhat.com>
Patchwork-id: 90361
O-Subject: [RHEL7.1 PATCH 169/281] i40e: Allow RSS table entry range and GPS to be any number, not necessarily power of 2
Bugzilla: 1091126
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Anjali Singhai Jain <anjali.singhai@intel.com>

We tell the HW upper boundary of power of 2 in VSI config,
but the HW does not restrict us to use just power of 2 GPS in
case of RSS as long as we are not sharing the RSS table with
another VSI (VMDq). We at present are not doing RSS in VMDq
VSI.
If we were to enable that and if the system had CPU count which
was not power 2, the VMDq VSIs will see a little skewed distribution.

Change-ID: I3ea797ce9065a3ca4fc4d04251bf195463410473
Signed-off-by: Anjali Singhai Jain <anjali.singhai@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit eb051afedf0ace9bf1356d5178f08ef00e6e3249)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>