[ethernet] i40e: Add PF reset when Malicious driver event for PF
Message-id: <1409658042-22689-198-git-send-email-sassmann@redhat.com>
Patchwork-id: 90391
O-Subject: [RHEL7.1 PATCH 197/281] i40e: Add PF reset when Malicious driver event for PF
Bugzilla: 1091126
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Neerav Parikh <neerav.parikh@intel.com>

As per the spec when the PF driver receives a Malicious driver event
the queue that caused the event is already stopped and it is expected
that the function that owns the queue will reset the queue.
In some cases it may not be possible to determine the queue and it is
suggested to reset the whole function.

This patch takes the later approach when the event is owned by the PF
that owns it.

Change-ID: I40f9764a6a5e068c0ef8438db00c5aa9c2c6c1c8
Signed-off-by: Neerav Parikh <neerav.parikh@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit df430b1240d3375d053230d1f943383d59f9267a)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>