[alsa] hda/sigmatel: Allow auto-switching for dock line-in of HP laptops
Message-id: <140972823573-198-patchset-jkysela@redhat.com>
Patchwork-id: 90898
O-Subject: [RHEL 7.1 PATCH 197/442] ALSA: hda/sigmatel - Allow auto-switching for dock line-in of HP laptops
Bugzilla: 1112200
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Many HP laptops with STAC codecs have a docking station port and BIOS
sets the pins for the input on the dock as a line in.  Because the
generic parser doesn't handle a line in pin as auto-switchable, this
resulted in the manual capture source selection on these laptops.

However, from the usability POV, the automatic switching is easier.
This patch adds the line_in_auto_switch hint in the fixup function for
these laptops.  Even if no dock port is present, this should be
harmless as the generic parser allows the auto-switching only in a
limited situation (all three pins are located in different
positions).

Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: e8b99a1dcb49b0d362b19a4831a00d85c76bd4b3)
rhbz#1112200
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>